Fixed tooltips opened via API on click closing immediately.
The global document click-to-close has been changed so it will no longer
be applied to tooltips opened via the API.

DisplayController has been updated to cancel and close any tooltips that
are in their close delay period when it is asked to do an immediate
open. This is to prevent close delay collisions causing unexpected
behavior in this scenario.

For issue #127.