Changed jQuery to be a required dependency in npm installs.
This promotes the "jquery" npm package from dev dependency to full
dependency. Since jQuery is absolutely required for PowerTip to
function this is probably the best choice to guarantee that jQuery
is present on any npm installs.

I decided against using a peer dependency because PowerTip will
actually `require('jquery')`, and npm >= 3.0 does not automatically
install peer dependencies.

For issue #134.