Changed core event hooking unit tests to use triggerHandler
This change is to support Internet Explorer in the unit tests.
Internet Explorer has a brief delay after invoking the focus and
blur events. This change bypasses the full eventing system in the
browser, and asks jQuery to fire the events it knows about.

This isn't ideal but it's either this or add a bunch of hacky async
code based on timeouts to the tests. In this case, I believe we can
assume that jQuery's event binding is well tested enough to assume
that it works correctly.