Fixed non-functional tooltip bug when repeatedly calling powerTip
This patch fixes two issues that would lead to tooltips not working
on an even number of repeated powerTip() invocations on the same
DOM elements.

      * The data attributes were being loaded before the destroy().
      * The TooltipController (and tooltip element creation) was being
        run before the destroy() call.

The first bug has been present since the 1.2.0. The second was
a side-effect of changing the destroy() method in 1.3.0 to remove
the tooltip element if all tooltip instances had been removed.

Reference commits:
  * 25ce6b1a9060886332cdfff8e9954ea730a5cbdf
  * f08fbecc486a98b8d9f44dc2e6acd8042e06446e

For issue #157.