Change isMouseEvent() to compare event type to mouse event list
The original check isn’t valid anymore. It seems that jQuery 3.x is
being “helpful” by normalizing all events to include pageX and
pageY properties, even if the event that triggered it doesn’t
supply that information. In jQuery versions 1 and 2 these
properties are undefined if there was no relevant data, but in
jQuery 3 they are set to `0`.

This patch changes the check to match the event type to a list of
known mouse events that are expected to supply valid pageX and
pageY data.

Part of issue #158.