Changed mouse-follow tips to use static placement on non-mouse event
This will force tooltips opened with the `followMouse` option to
use a static placement option with smart placement when they are
opened with an event that doesn't have mouse cursor position data.

With this change users navigating with keyboard navigation will
now be able to trigger these tooltips and have them show in a
reasonable position on the page. The change also fixes a bug where
focusing an element with a tooltip (e.g. clicking it) and returning
to the page from another browser tab was firing the focus event,
causing the tooltip to display at the last know mouse position.

For issue #147.